requirements.txt
diff-match-patch==20121119
Django==2.0.4
django-appconf==1.0.2
django-countries==5.3
django-crispy-forms==1.7.2
django-filter==1.1.0
django-imagekit==4.0.2
django-import-export==1.0.0
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
pilkit==2.0
Pillow==5.1.0
pytz==2018.4
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
dj-database-url==0.4.2
django-widget-tweaks
python-decouple